@@ -3,28 +3,32 @@ from ConfigSpace.hyperparameters.beta_integer import BetaIntegerHyperparameter
 from ConfigSpace.hyperparameters.categorical import CategoricalHyperparameter
 from ConfigSpace.hyperparameters.constant import Constant, UnParametrizedHyperparameter
 from ConfigSpace.hyperparameters.float_hyperparameter import FloatHyperparameter
-from ConfigSpace.hyperparameters.hyperparameter import Hyperparameter
+from ConfigSpace.hyperparameters.hyperparameter import (
+    Hyperparameter,
+    HyperparameterWithPrior,
+)
 from ConfigSpace.hyperparameters.integer_hyperparameter import IntegerHyperparameter
 from ConfigSpace.hyperparameters.normal_float import NormalFloatHyperparameter
 from ConfigSpace.hyperparameters.normal_integer import NormalIntegerHyperparameter
-from ConfigSpace.hyperparameters.numerical import NumericalHyperparameter
+from ConfigSpace.hyperparameters.numerical_hyperparameter import NumericalHyperparameter
 from ConfigSpace.hyperparameters.ordinal import OrdinalHyperparameter
 from ConfigSpace.hyperparameters.uniform_float import UniformFloatHyperparameter
 from ConfigSpace.hyperparameters.uniform_integer import UniformIntegerHyperparameter
 
 __all__ = [
-    "Hyperparameter",
-    "Constant",
-    "UnParametrizedHyperparameter",
-    "OrdinalHyperparameter",
-    "CategoricalHyperparameter",
-    "NumericalHyperparameter",
-    "FloatHyperparameter",
-    "IntegerHyperparameter",
-    "UniformFloatHyperparameter",
-    "UniformIntegerHyperparameter",
-    "NormalFloatHyperparameter",
-    "NormalIntegerHyperparameter",
     "BetaFloatHyperparameter",
     "BetaIntegerHyperparameter",
+    "CategoricalHyperparameter",
+    "Constant",
+    "FloatHyperparameter",
+    "Hyperparameter",
+    "HyperparameterWithPrior",
+    "IntegerHyperparameter",
+    "NormalFloatHyperparameter",
+    "NormalIntegerHyperparameter",
+    "NumericalHyperparameter",
+    "OrdinalHyperparameter",
+    "UniformFloatHyperparameter",
+    "UniformIntegerHyperparameter",
+    "UnParametrizedHyperparameter",
 ]
