@@ -1,101 +1,19 @@
 from __future__ import annotations
 
+from dataclasses import dataclass
+
 import numpy as np
 
-from ConfigSpace.hyperparameters.numerical import NumericalHyperparameter
+from ConfigSpace.hyperparameters.hyperparameter import Hyperparameter
+from ConfigSpace.hyperparameters.numerical_hyperparameter import NumericalHyperparameter
 
 
-class IntegerHyperparameter(NumericalHyperparameter):
-    def __init__(self, name: str, default_value: int, meta: dict | None = None) -> None:
-        super().__init__(name, default_value, meta)
-
-    def is_legal(self, value: int) -> bool:
-        raise NotImplementedError
-
-    def is_legal_vector(self, value) -> int:
-        raise NotImplementedError
-
-    def check_default(self, default_value) -> int:
-        raise NotImplementedError
-
-    def check_int(self, parameter: int, name: str) -> int:
-        if abs(int(parameter) - parameter) > 0.00000001 and type(parameter) is not int:
-            raise ValueError(
-                f"For the Integer parameter {name}, the value must be "
-                f"an Integer, too. Right now it is a {type(parameter)} with value"
-                f" {parameter!s}.",
-            )
-        return int(parameter)
-
-    def _transform(
-        self,
-        vector: np.ndarray | float | int,
-    ) -> np.ndarray | float | int | None:
-        try:
-            if isinstance(vector, np.ndarray):
-                return self._transform_vector(vector)
-            return self._transform_scalar(vector)
-        except ValueError:
-            return None
-
-    def _transform_scalar(self, scalar: float) -> float:
-        raise NotImplementedError()
-
-    def _transform_vector(self, vector: np.ndarray) -> np.ndarray:
-        raise NotImplementedError()
-
-    def pdf(self, vector: np.ndarray) -> np.ndarray:
-        """
-        Computes the probability density function of the hyperparameter in
-        the hyperparameter space (the one specified by the user).
-        For each hyperparameter type, there is also a method _pdf which
-        operates on the transformed (and possibly normalized) hyperparameter
-        space. Only legal values return a positive probability density,
-        otherwise zero.
-
-        Parameters
-        ----------
-        vector: np.ndarray
-            the (N, ) vector of inputs for which the probability density
-            function is to be computed.
-
-        Returns
-        -------
-        np.ndarray(N, )
-            Probability density values of the input vector
-        """
-        if vector.ndim != 1:
-            raise ValueError("Method pdf expects a one-dimensional numpy array")
-        is_integer = (np.round(vector) == vector).astype(int)
-        vector = self._inverse_transform(vector)
-        return self._pdf(vector) * is_integer
-
-    def _pdf(self, vector: np.ndarray) -> np.ndarray:
-        """
-        Computes the probability density function of the parameter in
-        the transformed (and possibly normalized, depends on the parameter
-        type) space. As such, one never has to worry about log-normal
-        distributions, only normal distributions (as the inverse_transform
-        in the pdf method handles these). Optimally, an IntegerHyperparameter
-        should have a corresponding float, which can be utlized for the calls
-        to the probability density function (see e.g. NormalIntegerHyperparameter).
-
-        Parameters
-        ----------
-        vector: np.ndarray
-            the (N, ) vector of inputs for which the probability density
-            function is to be computed.
-
-        Returns
-        -------
-        np.ndarray(N, )
-            Probability density values of the input vector
-        """
-        raise NotImplementedError()
-
-    def get_max_density(self) -> float:
-        """
-        Returns the maximal density on the pdf for the parameter (so not
-        the mode, but the value of the pdf on the mode).
-        """
-        raise NotImplementedError()
+@dataclass(init=False)
+class IntegerHyperparameter(
+    Hyperparameter[np.int64, np.float64],
+    NumericalHyperparameter[np.int64],
+):
+    def _neighborhood_size(self, value: np.int64 | None) -> int:
+        if value is None or self.lower <= value <= self.upper:
+            return int(self.size)
+        return int(self.size) - 1
